% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_brightfield_proj.R
\name{add_brightfield_proj}
\alias{add_brightfield_proj}
\title{Add details for a brightfield projection to the metadata object}
\usage{
add_brightfield_proj(file_f, p = "p16", flatfield_dir = "flatfield",
  brightfield_channel = "ch2", fk = "fk1fl1",
  projection_tag = "_projection", image_tag = ".tiff")
}
\arguments{
\item{name}{}
}
\description{
Add details for a brightfield projection to the metadata object
}
