% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_flourescent_proj.R
\name{add_flourescent_proj}
\alias{add_flourescent_proj}
\title{Add details for a feature extraction of multi-channel_n images to metadata}
\usage{
add_flourescent_proj(file_f, flatfield_dir = "flatfield",
  channel_n = c("ch3", "ch4"), projection_tag = "maxproject.tiff")
}
\arguments{
\item{name}{}
}
\description{
Add details for a feature extraction of multi-channel_n images to metadata
}
