% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_feature_data.R
\name{collect_feature_data}
\alias{collect_feature_data}
\title{Collect Feature data after image analysis with cellprofiler}
\usage{
collect_feature_data(path_data, channel_of_interest = "ch1",
  measurment_of_interest = "measurment_IdentifySecondaryObjects.csv")
}
\arguments{
\item{measurment_of_interest}{}
}
\description{
Collect Feature data after image analysis with cellprofiler
}
