% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_feature_data_db.R
\name{collect_feature_data_db}
\alias{collect_feature_data_db}
\title{Collect Feature data after image analysis with cellprofiler}
\usage{
collect_feature_data_db(path_data,
  database = "/home/ubuntu/bucket/data/cellevent_tmrm_cellprofiler.db",
  channel_of_interest = "ch1",
  measurment_of_interest = "measurment_IdentifySecondaryObjects.csv")
}
\arguments{
\item{measurment_of_interest}{}
}
\description{
Collect Feature data after image analysis with cellprofiler
}
