% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_brightfield_metadata_new.R
\name{create_brightfield_metadata_new}
\alias{create_brightfield_metadata_new}
\title{Title}
\usage{
create_brightfield_metadata_new(path = getwd(), ch_number = "ch1",
  interactive = TRUE, interactive_path = "~/bucket/metadata",
  name = "brightfield", bucket_mode = FALSE)
}
\arguments{
\item{name}{}
}
\description{
Title
}
