% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_flatfield_metadata_split.R
\name{create_flatfield_metadata_split}
\alias{create_flatfield_metadata_split}
\title{Title}
\usage{
create_flatfield_metadata_split(path = paste0(getwd(), "/"), ch_number,
  force = FALSE, name = "bf", range = c(1, 16))
}
\arguments{
\item{name}{}
}
\description{
Title
}
