% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_phase_metadata.R
\name{create_phase_metadata}
\alias{create_phase_metadata}
\title{Title}
\usage{
create_phase_metadata(path = getwd(), ch_number = "ch1",
  interactive = TRUE, interactive_path = "~/metadata", name = "pc",
  bucket_mode = TRUE)
}
\arguments{
\item{name}{}
}
\description{
Title
}
