% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dump_database.R
\name{dump_database}
\alias{dump_database}
\title{Dump a dataframe into a database}
\usage{
dump_database(x, name_of_table = "cells", pool = pool)
}
\arguments{
\item{measurment_of_interest}{}
}
\description{
Dump a dataframe into a database
}
