% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_filelist.R
\name{extract_filelist}
\alias{extract_filelist}
\title{Return tibble describing each image file with well, field etc. in a given directory}
\usage{
extract_filelist(path = getwd())
}
\arguments{
\item{path}{the path of the directry to screen. The path is not checked recursively}
}
\value{
a tibble with columns: file_name, n_zst, well:zst, l_row, channel, file_path, parent
}
\description{
Return tibble describing each image file with well, field etc. in a given directory
}
\examples{

}
