% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_group.R
\name{generate_group}
\alias{generate_group}
\title{Generate grouping function call}
\usage{
generate_group(plate_name, channel_n)
}
\arguments{
\item{channel_n}{}
}
\description{
Generate grouping function call
}
