% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_jobs_bash.R
\name{group_jobs_bash}
\alias{group_jobs_bash}
\title{Creating a shell script that submits a whole collection of Jobs for job files that match a given pattern and well range}
\usage{
group_jobs_bash(path_base, name, letter_row_interval, number_col_interval,
  python_call = "python ~/Distributed-CellProfiler/run.py submitJob ")
}
\arguments{
\item{path_base}{the path to the directory containing the job files}

\item{letter_row_interval}{a vector of the form "c(1:3)" that will be translated to row letters}

\item{number_col_interval}{a vector of the form "c(1:3)" that will be translated to col numbers}

\item{python_call}{a python call that will be created}

\item{pattern}{a string that has to be contained in the job files of interest}
}
\description{
Creating a shell script that submits a whole collection of Jobs for job files that match a given pattern and well range
}
