% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_data_file.R
\name{link_data_file}
\alias{link_data_file}
\title{Link metadata path into JSON job object}
\usage{
link_data_file(metadata_split_path, json_template)
}
\arguments{
\item{name}{}
}
\description{
Link metadata path into JSON job object
}
