% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_json_metadata.R
\name{link_json_metadata}
\alias{link_json_metadata}
\title{Linking JSON job file to metadata object}
\usage{
link_json_metadata(json_path, metadata_split_path,
  path = "~/bucket/metadata/")
}
\arguments{
\item{name}{}
}
\description{
Linking JSON job file to metadata object
}
