% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_046703learningdata.R
\name{pull_046703learningdata}
\alias{pull_046703learningdata}
\title{Prepare the learning dataset for 046703}
\usage{
pull_046703learningdata()
}
\arguments{
\item{name}{}
}
\description{
Prepare the learning dataset for 046703
}
\examples{

}
