% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_metadata.R
\name{write_metadata}
\alias{write_metadata}
\title{Title}
\usage{
write_metadata(df, path, interactive, interactive_path = "~/metadata",
  name = "pc")
}
\arguments{
\item{name}{}
}
\description{
Title
}
