% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_flatfield_metadata.R
\name{create_flatfield_metadata}
\alias{create_flatfield_metadata}
\title{Title}
\usage{
create_flatfield_metadata(path = paste0(getwd(), "/"), ch_number,
  force = FALSE, name = "bf", interactive = TRUE,
  interactive_path = "~/bucket/metadata", bucket_mode = FALSE)
}
\arguments{
\item{name}{}
}
\description{
Title
}
