% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_data_file.R
\name{link_data_file}
\alias{link_data_file}
\title{Fill a JSON template with a specific metadata path and metadata groupings}
\usage{
link_data_file(metadata_split_path, json_template, json_path, path_base)
}
\arguments{
\item{name}{}
}
\description{
Fill a JSON template with a specific metadata path and metadata groupings
}
